+++ app/signin/page.tsx
@@ -80,7 +80,7 @@ const Login: React.FC = () => {
                         onChange={handlePasswordChange}
                     />
                 </div>
-                <Link href="/map-page">
+                <Link href="/about">
                     <button className="main_button center h3 padding-all" onClick={handleLogin}>
                         Login
                     </button>
